fix(ios): address MIM-178 PR review findings
