+++ ios/MimiRemote/Sources/Features/Projects/WorkspaceStripPresentation.swift
@@ -20,8 +20,8 @@ enum WorkspaceStripLayout {
     /// 行末「添加工作区」虚线胶囊的可见尺寸。刻意比 `chipHeight` 小一圈：
     /// 它是次级动作，不该和承载工作区身份的项目胶囊等重。命中区仍由外层撑满 44pt。
     static let addChipVisualSize: CGFloat = 34
-    /// Runtime 筛选器并入胶囊行所需的宽度，单位是**扣掉设备入口之后的整行容器宽**，
-    /// 不是设备屏宽——容器宽已经减掉了两侧 `horizontalPadding`。
+    /// Runtime 筛选器并入胶囊行所需的宽度，单位是扣掉两侧 `horizontalPadding`
+    /// 和设备入口预算后的实际内容宽。
     ///
     /// 预算：筛选器 ~158 + 分隔 ~9 + 添加胶囊 52 + 选中胶囊 ~95 + 至少一屏可滑的胶囊余量 ~150。
     /// 结果：iPad mini 竖屏（744 屏宽 / 696 容器）与更宽的 iPad 进入合并态；
@@ -30,6 +30,23 @@ enum WorkspaceStripLayout {
 
     static func minimumContentWidth(viewportWidth: CGFloat) -> CGFloat {
         max(0, viewportWidth - horizontalPadding * 2)
+    }
+
+    /// 统一决定 Runtime 是否进入胶囊行，避免 Runtime 布局和行内新建入口分别判断。
+    /// `viewportWidth` 是应用了内边距的外层容器宽，不是设备屏宽。
+    static func usesInlineRuntimePicker(
+        viewportWidth: CGFloat,
+        showsHostSwitcherInStrip: Bool,
+        hasBottomTabBar: Bool
+    ) -> Bool {
+        // 底部 Tab 栏时新建按钮必须留在筛选行，横屏 iPhone 也不能切到 inline 布局。
+        guard !hasBottomTabBar else { return false }
+
+        let hostSwitcherBudget = showsHostSwitcherInStrip
+            ? WorkbenchChromeIconMetrics.minimumHitTarget + chipSpacing
+            : 0
+        let availableContentWidth = minimumContentWidth(viewportWidth: viewportWidth)
+        return availableContentWidth - hostSwitcherBudget >= inlineRuntimePickerMinimumWidth
     }
 
     /// 使用胶囊滚动区的真实宽度，而不是设备宽度。浮动侧栏会持续改变详情区宽度，
